Split resolvers into separate files
